refactor(auth): update related code after splitting AuthService and updating interfaces
